services/llm-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
torch==2.5.1
transformers==4.46.2
sentence-transformers==3.3.0
faiss-cpu==1.9.0
numpy==1.26.4
python-multipart==0.0.12
